+++ src/main/java/Parser.java
@@ -1,7 +1,7 @@
 import Task.Deadline;
+import Task.Event;
 import Task.Task;
 import Task.Todo;
-import Task.Event;
 
 import java.text.ParseException;
 import java.util.ArrayList;
