+++ src/main/test/DukeTest/TodoTest.java
@@ -1,5 +1,6 @@
-package Task;
+package DukeTest;
 
+import Task.Todo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
